fixup! Bluetooth timeout feature

+++ src/com/android/settings/bluetooth/BluetoothTimeoutPreferenceController.java
@@ -73,8 +73,8 @@ public class BluetoothTimeoutPreferenceController extends BasePreferenceControll
     @Override
     public boolean onPreferenceChange(Preference preference, Object newValue) {
         try {
-            int value = Integer.parseInt((String) newValue);
-            LineageSettings.Global.putInt(mContext.getContentResolver(),
+            long value = Long.parseLong((String) newValue);
+            LineageSettings.Global.putLong(mContext.getContentResolver(),
                     LineageSettings.Global.BLUETOOTH_OFF_TIMEOUT, value);
             updateTimeoutPreferenceDescription((ListPreference) preference, value);
         } catch (NumberFormatException e) {
